[Feat] 강의평가 상세 화면 구현 및 서버 연동

+++ SUWIKI/SUWIKI/Data/DTO/Lecture/Response/DTO+DetailLectureResponse.swift
@@ -49,9 +49,9 @@ extension DTO.DetailLectureResponse {
                       lectureSatisfactionAvg: Double(lectureSatisfactionAvg),
                       lectureHoneyAvg: Double(lectureHoneyAvg),
                       lectureLearningAvg: Double(lectureLearningAvg),
-                      lectureDifficultyAvg: Double(lectureDifficultyAvg),
-                      lectureTeamAvg: Double(lectureTeamAvg),
-                      lectureHomeworkAvg: Double(lectureHomeworkAvg))
+                      lectureDifficultyAvg: Int(lectureDifficultyAvg),
+                      lectureTeamAvg: Int(lectureTeamAvg),
+                      lectureHomeworkAvg: Int(lectureHomeworkAvg))
     }
 }
